@@ -10,7 +10,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_svg: ^2.0.7
-  http: ^0.13.3
+  http: ^1.1.0
   carousel_slider: ^4.0.0
   intl: ^0.18.0
   flutter_secure_storage: ^9.0.0
@@ -23,14 +23,14 @@ dependencies:
   image_picker: ^1.0.4
   firebase_messaging: ^14.7.9
   firebase_core: ^2.24.2
-  webview_flutter: ^2.8.0
-  flutter_recaptcha_v3: ^0.1.3
+  webview_flutter: ^4.4.2
+ 
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
 flutter:
   uses-material-design: true
